Handle function calls with unresolved variables
Certain function calls like `merge` will not be executed if their
arguments include unresolved variables such as data blocks that we do
not execute.

If a function call is made and it returns `nil` we now do some custom
processing to run the functions with all their arguments except the
unresolved values.