fix: Handle function calls with unresolved variables (#238)
Certain function calls like `merge` will not be executed if their
arguments include unresolved variables such as data blocks that we do
not execute.

If a function call is made and it returns `nil` we now do some custom
processing to run the functions with all their arguments except the
unresolved values.

Co-authored-by: AJ Kerrigan <aj@speckledmonkey.com>