fix: unresolved JSON strings aren't valid JSON
When we have a JSON string (like container definitions) if it is
unresolved we are inserting a reference string in there to what was
unresolved but its not valid JSON so if anyone uses `fromjson()` on it
it will fail.

This detects JSON strings and formats it properly while maintaining
the reference.